Move EndpointBuilder out and do some other cleanups
