+++ pilot/pkg/networking/util/util.go
@@ -500,7 +500,7 @@ func MergeAnyWithAny(dst *any.Any, src *any.Any) (*any.Any, error) {
 }
 
 // BuildLbEndpointMetadata adds metadata values to a lb endpoint
-func BuildLbEndpointMetadata(network string, tlsMode string, push *model.PushContext) *core.Metadata {
+func BuildLbEndpointMetadata(network string, tlsMode string) *core.Metadata {
 	if network == "" && tlsMode == model.DisabledTLSModeLabel {
 		return nil
 	}
