+++ pilot/pkg/xds/eds.go
@@ -17,74 +17,21 @@ package xds
 import (
 	"time"
 
-	core "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
 	endpoint "github.com/envoyproxy/go-control-plane/envoy/config/endpoint/v3"
 	discovery "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v3"
 	"github.com/golang/protobuf/ptypes/any"
-	"github.com/golang/protobuf/ptypes/wrappers"
-
 	networkingapi "istio.io/api/networking/v1alpha3"
+
 	"istio.io/istio/pilot/pkg/model"
 	networking "istio.io/istio/pilot/pkg/networking/core/v1alpha3"
 	"istio.io/istio/pilot/pkg/networking/core/v1alpha3/loadbalancer"
 	"istio.io/istio/pilot/pkg/networking/util"
 	"istio.io/istio/pilot/pkg/util/sets"
 	v3 "istio.io/istio/pilot/pkg/xds/v3"
-	"istio.io/istio/pkg/config/host"
 	"istio.io/istio/pkg/config/labels"
 	"istio.io/istio/pkg/config/protocol"
 	"istio.io/istio/pkg/config/schema/gvk"
 )
-
-// EDS returns the list of endpoints (IP:port and in future labels) associated with a real
-// service or a subset of a service, selected using labels.
-//
-// The source of info is a list of service registries.
-//
-// Primary event is an endpoint creation/deletion. Once the event is fired, EDS needs to
-// find the list of services associated with the endpoint.
-//
-// In case of k8s, Endpoints event is fired when the endpoints are added to service - typically
-// after readiness check. At that point we have the 'real' Service. The Endpoint includes a list
-// of port numbers and names.
-//
-// For the subset case, the Pod referenced in the Endpoint must be looked up, and pod checked
-// for labels.
-//
-// In addition, ExternalEndpoint includes IPs and labels directly and can be directly processed.
-//
-// TODO: for selector-less services (mesh expansion), skip pod processing
-// TODO: optimize the code path for ExternalEndpoint, no additional processing needed
-// TODO: if a service doesn't have split traffic - we can also skip pod and label processing
-// TODO: efficient label processing. In alpha3, the destination policies are set per service, so
-// we may only need to search in a small list.
-
-// buildEnvoyLbEndpoint packs the endpoint based on istio info.
-func buildEnvoyLbEndpoint(e *model.IstioEndpoint, push *model.PushContext) *endpoint.LbEndpoint {
-	addr := util.BuildAddress(e.Address, e.EndpointPort)
-
-	epWeight := e.LbWeight
-	if epWeight == 0 {
-		epWeight = 1
-	}
-	ep := &endpoint.LbEndpoint{
-		LoadBalancingWeight: &wrappers.UInt32Value{
-			Value: epWeight,
-		},
-		HostIdentifier: &endpoint.LbEndpoint_Endpoint{
-			Endpoint: &endpoint.Endpoint{
-				Address: addr,
-			},
-		},
-	}
-
-	// Istio telemetry depends on the metadata value being set for endpoints in the mesh.
-	// Istio endpoint level tls transport socket configuration depends on this logic
-	// Do not remove
-	ep.Metadata = util.BuildLbEndpointMetadata(e.Network, e.TLSMode, push)
-
-	return ep
-}
 
 // UpdateServiceShards will list the endpoints and create the shards.
 // This is used to reconcile and to support non-k8s registries (until they migrate).
@@ -122,8 +69,7 @@ func (s *DiscoveryServer) UpdateServiceShards(push *model.PushContext) error {
 				}
 			}
 
-			// TODO(nmittler): Should we get the cluster from the endpoints instead? May require organizing endpoints by cluster first.
-			s.edsUpdate(registry.Cluster(), string(svc.Hostname), svc.Attributes.Namespace, endpoints)
+			s.edsCacheUpdate(registry.Cluster(), string(svc.Hostname), svc.Attributes.Namespace, endpoints)
 		}
 	}
 
@@ -136,20 +82,10 @@ func (s *DiscoveryServer) SvcUpdate(cluster, hostname string, namespace string, 
 	// prevent memory leaks.
 	if event == model.EventDelete {
 		inboundServiceDeletes.Increment()
-		s.mutex.Lock()
-		defer s.mutex.Unlock()
 		s.deleteService(cluster, hostname, namespace)
 	} else {
 		inboundServiceUpdates.Increment()
 	}
-}
-
-// Update clusters for an incremental EDS push, and initiate the push.
-// Only clusters that changed are updated/pushed.
-func (s *DiscoveryServer) edsIncremental(version string, req *model.PushRequest) {
-	adsLog.Infof("XDS:EDSInc Pushing:%s Services:%v ConnectedEndpoints:%d",
-		version, model.ConfigNamesOfKind(req.ConfigsUpdated, gvk.ServiceEntry), s.adsClientCount())
-	s.startPush(req)
 }
 
 // EDSUpdate computes destination address membership across all clusters and networks.
@@ -160,8 +96,8 @@ func (s *DiscoveryServer) edsIncremental(version string, req *model.PushRequest)
 func (s *DiscoveryServer) EDSUpdate(clusterID, serviceName string, namespace string,
 	istioEndpoints []*model.IstioEndpoint) {
 	inboundEDSUpdates.Increment()
-	// Update the eds data structures
-	fp := s.edsUpdate(clusterID, serviceName, namespace, istioEndpoints)
+	// Update the endpoint shards
+	fp := s.edsCacheUpdate(clusterID, serviceName, namespace, istioEndpoints)
 	// Trigger a push
 	s.ConfigUpdate(&model.PushRequest{
 		Full: fp,
@@ -174,7 +110,7 @@ func (s *DiscoveryServer) EDSUpdate(clusterID, serviceName string, namespace str
 	})
 }
 
-// EDSUpdate computes destination address membership across all clusters and networks.
+// EDSCacheUpdate computes destination address membership across all clusters and networks.
 // This is the main method implementing EDS.
 // It replaces InstancesByPort in model - instead of iterating over all endpoints it uses
 // the hostname-keyed map. And it avoids the conversion from Endpoint to ServiceEntry to envoy
@@ -184,28 +120,32 @@ func (s *DiscoveryServer) EDSUpdate(clusterID, serviceName string, namespace str
 func (s *DiscoveryServer) EDSCacheUpdate(clusterID, serviceName string, namespace string,
 	istioEndpoints []*model.IstioEndpoint) {
 	inboundEDSUpdates.Increment()
-	// Update the eds data structures
-	s.edsUpdate(clusterID, serviceName, namespace, istioEndpoints)
+	// Update the endpoint shards
+	s.edsCacheUpdate(clusterID, serviceName, namespace, istioEndpoints)
 }
 
-// edsUpdate updates EndpointShards data by clusterID, serviceName, IstioEndpoints.
+// edsCacheUpdate updates EndpointShards data by clusterID, hostname, IstioEndpoints.
 // It also tracks the changes to ServiceAccounts. It returns whether a full push
 // is needed or incremental push is sufficient.
-func (s *DiscoveryServer) edsUpdate(clusterID, serviceName string, namespace string,
+func (s *DiscoveryServer) edsCacheUpdate(clusterID, hostname string, namespace string,
 	istioEndpoints []*model.IstioEndpoint) bool {
 	if len(istioEndpoints) == 0 {
-		s.handleEmptyEndpoints(clusterID, serviceName, namespace)
+		// Should delete the service EndpointShards when endpoints become zero to prevent memory leak,
+		// but we should not do not delete the keys from EndpointShardsByService map - that will trigger
+		// unnecessary full push which can become a real problem if a pod is in crashloop and thus endpoints
+		// flip flopping between 1 and 0.
+		s.deleteEndpointShards(clusterID, hostname, namespace)
+		adsLog.Infof("Incremental push, service %s has no endpoints", hostname)
 		return false
 	}
 
 	fullPush := false
 
 	// Find endpoint shard for this service, if it is available - otherwise create a new one.
-	ep, created := s.getOrCreateEndpointShard(serviceName, namespace)
-
+	ep, created := s.getOrCreateEndpointShard(hostname, namespace)
 	// If we create a new endpoint shard, that means we have not seen the service earlier. We should do a full push.
 	if created {
-		adsLog.Infof("Full push, new service %s", serviceName)
+		adsLog.Infof("Full push, new service %s", hostname)
 		fullPush = true
 	}
 
@@ -219,32 +159,17 @@ func (s *DiscoveryServer) edsUpdate(clusterID, serviceName string, namespace str
 
 	ep.mutex.Lock()
 	// For existing endpoints, we need to do full push if service accounts change.
-	if !created {
-		if !serviceAccounts.Equals(ep.ServiceAccounts) {
-			adsLog.Debugf("Updating service accounts now, svc %v, before service account %v, after %v",
-				serviceName, ep.ServiceAccounts, serviceAccounts)
-			adsLog.Infof("Full push, service accounts changed, %v", serviceName)
-			fullPush = true
-		}
+	if !fullPush && !serviceAccounts.Equals(ep.ServiceAccounts) {
+		adsLog.Debugf("Updating service accounts now, svc %v, before service account %v, after %v",
+			hostname, ep.ServiceAccounts, serviceAccounts)
+		adsLog.Infof("Full push, service accounts changed, %v", hostname)
+		fullPush = true
 	}
 	ep.Shards[clusterID] = istioEndpoints
 	ep.ServiceAccounts = serviceAccounts
 	ep.mutex.Unlock()
 
 	return fullPush
-}
-
-func (s *DiscoveryServer) handleEmptyEndpoints(clusterID, serviceName, namespace string) {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-	// Should delete the service EndpointShards when endpoints become zero to prevent memory leak,
-	// but we should not do not delete the keys from EndpointShardsByService map - that will trigger
-	// unnecessary full push which can become a real problem if a pod is in crashloop and thus endpoints
-	// flip flopping between 1 and 0.
-	if s.EndpointShardsByService[serviceName][namespace] != nil {
-		s.deleteEndpointShards(clusterID, serviceName, namespace)
-		adsLog.Infof("Incremental push, service %s has no endpoints", serviceName)
-	}
 }
 
 func (s *DiscoveryServer) getOrCreateEndpointShard(serviceName, namespace string) (*EndpointShards, bool) {
@@ -270,7 +195,10 @@ func (s *DiscoveryServer) getOrCreateEndpointShard(serviceName, namespace string
 // deleteEndpointShards deletes matching endpoint shards from EndpointShardsByService map. This is called when
 // endpoints are deleted.
 func (s *DiscoveryServer) deleteEndpointShards(cluster, serviceName, namespace string) {
-	if s.EndpointShardsByService[serviceName][namespace] != nil {
+	s.mutex.Lock()
+	defer s.mutex.Unlock()
+	if s.EndpointShardsByService[serviceName] != nil &&
+		s.EndpointShardsByService[serviceName][namespace] != nil {
 		s.EndpointShardsByService[serviceName][namespace].mutex.Lock()
 		delete(s.EndpointShardsByService[serviceName][namespace].Shards, cluster)
 		s.EndpointShardsByService[serviceName][namespace].mutex.Unlock()
@@ -280,12 +208,18 @@ func (s *DiscoveryServer) deleteEndpointShards(cluster, serviceName, namespace s
 // deleteService deletes all service related references from EndpointShardsByService. This is called
 // when a service is deleted.
 func (s *DiscoveryServer) deleteService(cluster, serviceName, namespace string) {
-	if s.EndpointShardsByService[serviceName][namespace] != nil {
+	s.mutex.Lock()
+	defer s.mutex.Unlock()
+
+	if s.EndpointShardsByService[serviceName] != nil &&
+		s.EndpointShardsByService[serviceName][namespace] != nil {
+
 		s.EndpointShardsByService[serviceName][namespace].mutex.Lock()
 		delete(s.EndpointShardsByService[serviceName][namespace].Shards, cluster)
-		svcShards := len(s.EndpointShardsByService[serviceName][namespace].Shards)
+		shards := len(s.EndpointShardsByService[serviceName][namespace].Shards)
 		s.EndpointShardsByService[serviceName][namespace].mutex.Unlock()
-		if svcShards == 0 {
+
+		if shards == 0 {
 			delete(s.EndpointShardsByService[serviceName], namespace)
 		}
 		if len(s.EndpointShardsByService[serviceName]) == 0 {
@@ -294,36 +228,10 @@ func (s *DiscoveryServer) deleteService(cluster, serviceName, namespace string) 
 	}
 }
 
-// SubsetToLabels returns the labels associated with a subset of a given service.
-func subsetToLabels(dr *model.Config, subsetName string) labels.Collection {
-	// empty subset
-	if subsetName == "" {
-		return nil
-	}
-
-	if dr == nil {
-		return nil
-	}
-
-	rule := dr.Spec.(*networkingapi.DestinationRule)
-	for _, subset := range rule.Subsets {
-		if subset.Name == subsetName {
-			if len(subset.Labels) == 0 {
-				return nil
-			}
-			return []labels.Instance{subset.Labels}
-		}
-	}
-
-	return nil
-}
-
 // loadAssignmentsForCluster return the endpoints for a cluster
 // Initial implementation is computing the endpoints on the flight - caching will be added as needed, based on
 // perf tests.
 func (s *DiscoveryServer) loadAssignmentsForCluster(b EndpointBuilder) *endpoint.ClusterLoadAssignment {
-	subsetLabels := subsetToLabels(b.destinationRule, b.subsetName)
-
 	if b.service == nil {
 		// Shouldn't happen here
 		adsLog.Debugf("can not find the service for cluster %s", b.clusterName)
@@ -338,7 +246,7 @@ func (s *DiscoveryServer) loadAssignmentsForCluster(b EndpointBuilder) *endpoint
 	// automatically.
 	// Gateways use EDS for Passthrough cluster. So we should allow Passthrough here.
 	if b.service.Resolution == model.DNSLB {
-		adsLog.Infof("cluster %s in  eds cluster, but its resolution now is updated to %v, skipping it.", b.clusterName, b.service.Resolution)
+		adsLog.Infof("cluster %s in eds cluster, but its resolution now is updated to %v, skipping it.", b.clusterName, b.service.Resolution)
 		return nil
 	}
 
@@ -349,9 +257,8 @@ func (s *DiscoveryServer) loadAssignmentsForCluster(b EndpointBuilder) *endpoint
 		return buildEmptyClusterLoadAssignment(b.clusterName)
 	}
 
-	// The service was never updated - do the full update
 	s.mutex.RLock()
-	se, f := s.EndpointShardsByService[string(b.hostname)][b.service.Attributes.Namespace]
+	epShards, f := s.EndpointShardsByService[string(b.hostname)][b.service.Attributes.Namespace]
 	s.mutex.RUnlock()
 	if !f {
 		// Shouldn't happen here
@@ -359,49 +266,12 @@ func (s *DiscoveryServer) loadAssignmentsForCluster(b EndpointBuilder) *endpoint
 		return buildEmptyClusterLoadAssignment(b.clusterName)
 	}
 
-	locEps := buildLocalityLbEndpointsFromShards(b, se, svcPort, subsetLabels)
+	locEps := b.buildLocalityLbEndpointsFromShards(epShards, svcPort)
 
 	return &endpoint.ClusterLoadAssignment{
 		ClusterName: b.clusterName,
 		Endpoints:   locEps,
 	}
-}
-
-type EndpointBuilder struct {
-	// These fields define the primary key for an endpoint, and can be used as a cache key
-	clusterName     string
-	network         string
-	clusterID       string
-	locality        *core.Locality
-	destinationRule *model.Config
-	service         *model.Service
-
-	// These fields are provided for convenience only
-	subsetName string
-	hostname   host.Name
-	port       int
-	push       *model.PushContext
-}
-
-func createEndpointBuilder(clusterName string, proxy *model.Proxy, push *model.PushContext) EndpointBuilder {
-	_, subsetName, hostname, port := model.ParseSubsetKey(clusterName)
-	svc := push.ServiceForHostname(proxy, hostname)
-
-	key := EndpointBuilder{
-		clusterName:     clusterName,
-		network:         proxy.Metadata.Network,
-		clusterID:       proxy.Metadata.ClusterID,
-		locality:        proxy.Locality,
-		service:         svc,
-		destinationRule: push.DestinationRule(proxy, svc),
-
-		push:       push,
-		subsetName: subsetName,
-		hostname:   hostname,
-		port:       port,
-	}
-
-	return key
 }
 
 func (s *DiscoveryServer) generateEndpoints(b EndpointBuilder) *endpoint.ClusterLoadAssignment {
@@ -413,13 +283,7 @@ func (s *DiscoveryServer) generateEndpoints(b EndpointBuilder) *endpoint.Cluster
 	// If networks are set (by default they aren't) apply the Split Horizon
 	// EDS filter on the endpoints
 	if b.push.Networks != nil && len(b.push.Networks.Networks) > 0 {
-		endpoints := EndpointsByNetworkFilter(b.push, b.network, l.Endpoints)
-		filteredCLA := &endpoint.ClusterLoadAssignment{
-			ClusterName: l.ClusterName,
-			Endpoints:   endpoints,
-			Policy:      l.Policy,
-		}
-		l = filteredCLA
+		l.Endpoints = EndpointsByNetworkFilter(b.push, b.network, l.Endpoints)
 	}
 
 	// If locality aware routing is enabled, prioritize endpoints or set their lb weight.
@@ -455,7 +319,7 @@ func (eds *EdsGenerator) Generate(proxy *model.Proxy, push *model.PushContext, w
 		if _, f := edsUpdatedServices[string(hostname)]; f {
 			continue
 		}
-		epb := createEndpointBuilder(clusterName, proxy, push)
+		epb := NewEndpointBuilder(clusterName, proxy, push)
 		l := eds.Server.generateEndpoints(epb)
 		if l == nil {
 			continue
@@ -487,7 +351,7 @@ func (s *DiscoveryServer) pushEds(push *model.PushContext, con *Connection, vers
 				continue
 			}
 		}
-		builder := createEndpointBuilder(clusterName, con.proxy, push)
+		builder := NewEndpointBuilder(clusterName, con.proxy, push)
 		l := s.generateEndpoints(builder)
 		if l == nil {
 			continue
@@ -522,11 +386,10 @@ func (s *DiscoveryServer) pushEds(push *model.PushContext, con *Connection, vers
 	return nil
 }
 
-func getOutlierDetectionAndLoadBalancerSettings(cfg *model.Config, portNumber int, subsetName string) (bool, *networkingapi.LoadBalancerSettings) {
-	if cfg == nil {
+func getOutlierDetectionAndLoadBalancerSettings(destinationRule *networkingapi.DestinationRule, portNumber int, subsetName string) (bool, *networkingapi.LoadBalancerSettings) {
+	if destinationRule == nil {
 		return false, nil
 	}
-	destinationRule := cfg.Spec.(*networkingapi.DestinationRule)
 	var outlierDetectionEnabled = false
 	var lbSettings *networkingapi.LoadBalancerSettings
 
@@ -565,76 +428,6 @@ func endpointDiscoveryResponse(loadAssignments []*endpoint.ClusterLoadAssignment
 	}
 
 	return out
-}
-
-// build LocalityLbEndpoints for a cluster from existing EndpointShards.
-func buildLocalityLbEndpointsFromShards(
-	b EndpointBuilder,
-	shards *EndpointShards,
-	svcPort *model.Port,
-	epLabels labels.Collection,
-) []*endpoint.LocalityLbEndpoints {
-	localityEpMap := make(map[string]*endpoint.LocalityLbEndpoints)
-
-	// Determine whether or not the target service is considered local to the cluster
-	// and should, therefore, not be accessed from outside the cluster.
-	isClusterLocal := b.push.IsClusterLocal(b.service)
-
-	shards.mutex.Lock()
-
-	// The shards are updated independently, now need to filter and merge
-	// for this cluster
-	for clusterID, endpoints := range shards.Shards {
-		// If the downstream service is configured as cluster-local, only include endpoints that
-		// reside in the same cluster.
-		if isClusterLocal && (clusterID != b.clusterID) {
-			continue
-		}
-
-		for _, ep := range endpoints {
-			if svcPort.Name != ep.ServicePortName {
-				continue
-			}
-			// Port labels
-			if !epLabels.HasSubsetOf(ep.Labels) {
-				continue
-			}
-
-			locLbEps, found := localityEpMap[ep.Locality.Label]
-			if !found {
-				locLbEps = &endpoint.LocalityLbEndpoints{
-					Locality:    util.ConvertLocality(ep.Locality.Label),
-					LbEndpoints: make([]*endpoint.LbEndpoint, 0, len(endpoints)),
-				}
-				localityEpMap[ep.Locality.Label] = locLbEps
-			}
-			if ep.EnvoyEndpoint == nil {
-				ep.EnvoyEndpoint = buildEnvoyLbEndpoint(ep, b.push)
-			}
-			locLbEps.LbEndpoints = append(locLbEps.LbEndpoints, ep.EnvoyEndpoint)
-
-		}
-	}
-
-	shards.mutex.Unlock()
-
-	locEps := make([]*endpoint.LocalityLbEndpoints, 0, len(localityEpMap))
-	for _, locLbEps := range localityEpMap {
-		var weight uint32
-		for _, ep := range locLbEps.LbEndpoints {
-			weight += ep.LoadBalancingWeight.GetValue()
-		}
-		locLbEps.LoadBalancingWeight = &wrappers.UInt32Value{
-			Value: weight,
-		}
-		locEps = append(locEps, locLbEps)
-	}
-
-	if len(locEps) == 0 {
-		b.push.AddMetric(model.ProxyStatusClusterNoInstances, b.clusterName, nil, "")
-	}
-
-	return locEps
 }
 
 // cluster with no endpoints
